Implement command q to manipulate points in groups.
This is needed because multiple threads are now writting and reading from the same memory,
and worse still is that memory is heap memory, and can be realloced, freed, ...

I didn't want to do this, but I don't see any other way. Alternative is spaggeti code.

Maybe with spaggeti I could get better performace, idk...