Makefile rework.
DRY Makefile.

I was thinking of removing makefile and only using zig to build this project.
Zig build is faster than makefile, it's more extendable, more human readable...

But, zig build changes from version to version. On My maching, I'm using zig 0.11.
Zig 0.10 or older can't run build.zig file. If I write the build script such that
it can be compiled with zig 0.10 then it can't be compiled with 0.11. This is
quite sad and I'm sad because now the best option for me is to keep 2 build
systems.