Bit of a refactor.
Made renering in such a way that all groups share one smol_mesh.
That change dictated the need for colors to be sent using in paramter, not as uniform.
Rearanged some things in all of the moduls.

Start of quad shader. Quads will be drawn by quad tree when quadarea/screenarea is smol.
This will drasticaly speedup plots that are not sorted on x axis. ( I hope )