A bit more of a refactor.
Header files should start with br_ prefix.
Source files should not.
prefix is used as to not use wrong header file if there are multiple inclue paths,
and to not have any confusion as to what header file is used.
Source files are not being searched in mutiple directories like header files are.
Also moved stuff like tests.h and benchmaks.h to src/misc folder to not cultter the
src folder.