resampling2.cpp -> resampling2.c
This is the only file that was cpp in raylib and headless version, but

it didn't had to be...

Cpp was usefull before when resampling was done using different

algorithm that was described nicely by templates.

Current implementation is not templatable, so there is no need for cpp..

:|

Also, implement more of a public interface.